+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [1.2.2] - 22nd of October, 2024
+
 ### Changed
 - Attempt device list download with max_items before falling back to full download
 
@@ -43,7 +45,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 - Initial release
 
-[Unreleased]: https://github.com/Syndace/slixmpp-omemo/compare/v1.2.1...HEAD
+[Unreleased]: https://github.com/Syndace/slixmpp-omemo/compare/v1.2.2...HEAD
+[1.2.2]: https://github.com/Syndace/slixmpp-omemo/compare/v1.2.1...v1.2.2
 [1.2.1]: https://github.com/Syndace/slixmpp-omemo/compare/v1.2.0...v1.2.1
 [1.2.0]: https://github.com/Syndace/slixmpp-omemo/compare/v1.1.1...v1.2.0
 [1.1.1]: https://github.com/Syndace/slixmpp-omemo/compare/v1.1.0...v1.1.1